User history, user groups, user autorization

+++ src/api/views/domain_views.py
@@ -8,7 +8,7 @@ from uuid import uuid4
 
 # Third-Party Libraries
 import boto3
-from flask import current_app, jsonify, request, send_file
+from flask import current_app, jsonify, request, send_file, g
 from flask.views import MethodView
 import requests
 from selenium import webdriver
@@ -29,6 +29,11 @@ from utils.categorization import (
 )
 from utils.two_captcha import two_captcha_api_key
 from utils.validator import validate_data
+from utils.user_profile import (
+    add_user_action, 
+    get_users_group_ids, 
+    user_can_access_domain,
+)
 
 category_manager = CategoryManager()
 proxy_manager = ProxyManager()
@@ -42,28 +47,42 @@ class DomainsView(MethodView):
 
     def get(self):
         """Get all domains."""
-        return jsonify(domain_manager.all(params=request.args))
+        add_user_action("Get Domains")
+
+        if g.is_admin:
+            response = domain_manager.all(params=request.args)
+        else:
+            groups = get_users_group_ids()
+            response = domain_manager.all(params= {
+                "application_id" : {'$in': groups }
+            })
+        
+        return jsonify(response)
 
     def post(self):
         """Create a new domain."""
-        data = validate_data(request.json, DomainSchema)
-        if domain_manager.get(filter_data={"name": data["name"]}):
-            return jsonify({"error": "Domain already exists."}), 400
-        caller_ref = str(uuid4())
-        resp = route53.create_hosted_zone(Name=data["name"], CallerReference=caller_ref)
-        domain_manager.save(
-            {
-                "name": data["name"],
-                "is_active": False,
-                "is_available": True,
-                "is_launching": False,
-                "is_delaunching": False,
-                "is_generating_template": False,
-                "route53": {"id": resp["HostedZone"]["Id"]},
-            }
-        )
-        return jsonify(resp["DelegationSet"]["NameServers"])
 
+        if g.is_admin:
+            data = validate_data(request.json, DomainSchema)
+            if domain_manager.get(filter_data={"name": data["name"]}):
+                return jsonify({"error": "Domain already exists."}), 400
+            caller_ref = str(uuid4())
+            resp = route53.create_hosted_zone(Name=data["name"], CallerReference=caller_ref)
+            domain_manager.save(
+                {
+                    "name": data["name"],
+                    "is_active": False,
+                    "is_available": True,
+                    "is_launching": False,
+                    "is_delaunching": False,
+                    "is_generating_template": False,
+                    "route53": {"id": resp["HostedZone"]["Id"]},
+                }
+            )
+            add_user_action(f"Created New Domain - {data['name']}")
+            return jsonify(resp["DelegationSet"]["NameServers"])
+        else:
+            return jsonify({"error": "User does not have admin rights, can not create a new domain"}), 400
 
 class DomainView(MethodView):
     """DomainView."""
@@ -71,14 +90,28 @@ class DomainView(MethodView):
     def get(self, domain_id):
         """Get Domain details."""
         domain = domain_manager.get(document_id=domain_id)
-        return jsonify(domain)
+        can_access = user_can_access_domain(domain)
+        if can_access:
+            add_user_action(f"View Domain - {domain['name']}")
+            if 'application_id' in domain:
+                application = application_manager.get(document_id=domain['application_id'])
+                domain['application_name'] = application['name']
+            return jsonify(domain)
+        else:
+            add_user_action(f"View Domain (ACCESS DENIED) - {domain['name']}")
+            return jsonify({"error": "User does not have permission to view this domain"}), 400
 
     def put(self, domain_id):
         """Update domain."""
         data = validate_data(request.json, DomainSchema)
+        domain = domain_manager.get(document_id=domain_id)
 
         if data.get("application"):
-            domain = domain_manager.get(document_id=domain_id)
+            can_access = user_can_access_domain(domain)
+            if not can_access:
+                add_user_action(f"Update Domain (ACCESS DENIED) - {domain['name']}")
+                return jsonify({"error": "User does not have permission to update this domain"}), 400
+
             application = application_manager.get(
                 filter_data={"name": data["application"]}
             )
@@ -92,11 +125,16 @@ class DomainView(MethodView):
                 }
             )
 
+        add_user_action(f"Update Domain - {domain['name']}")
         return jsonify(domain_manager.update(document_id=domain_id, data=data))
 
     def delete(self, domain_id):
         """Delete domain and hosted zone."""
         domain = domain_manager.get(document_id=domain_id)
+        can_access = user_can_access_domain(domain)
+        if not can_access:
+            add_user_action(f"Delete Domain (ACCESS DENIED) - {domain['name']}")
+            return jsonify({"error": "User does not have permission to delete this domain"}), 400
 
         if domain.get("is_active") and domain.get("redirects"):
             return jsonify(
@@ -111,6 +149,7 @@ class DomainView(MethodView):
             )
 
         route53.delete_hosted_zone(Id=domain["route53"]["id"])
+        add_user_action(f"Delete Domain - {domain['name']}")
         return jsonify(domain_manager.delete(domain["_id"]))
 
 
@@ -120,6 +159,11 @@ class DomainContentView(MethodView):
     def get(self, domain_id):
         """Download Domain."""
         domain = domain_manager.get(document_id=domain_id)
+        can_access = user_can_access_domain(domain)
+        if not can_access:
+            add_user_action(f"Download Domain (ACCESS DENIED) - {domain['name']}")
+            return jsonify({"error": "User does not have permission to download this domain"}), 400
+
 
         resp = requests.get(
             f"{STATIC_GEN_URL}/website/?category={domain['category']}&domain={domain['name']}",
@@ -134,6 +178,9 @@ class DomainContentView(MethodView):
         buffer.write(resp.content)
         buffer.seek(0)
 
+
+        add_user_action(f"Download Domain - {domain['name']}")
+
         return send_file(
             buffer,
             as_attachment=True,
@@ -145,6 +192,11 @@ class DomainContentView(MethodView):
         """Upload files and serve s3 site."""
         # Get domain data
         domain = domain_manager.get(document_id=domain_id)
+        can_access = user_can_access_domain(domain)
+        if not can_access:
+            add_user_action(f"Upload Domain (ACCESS DENIED) - {domain['name']}")
+            return jsonify({"error": "User does not have permission to upload to this domain"}), 400
+
 
         domain_name = domain["name"]
         category = request.args.get("category")
@@ -173,6 +225,9 @@ class DomainContentView(MethodView):
         # Remove temp files
         shutil.rmtree(f"tmp/{category}/", ignore_errors=True)
 
+
+        add_user_action(f"Upload Domain - {domain['name']}")
+
         return (
             jsonify(
                 domain_manager.update(
@@ -189,6 +244,11 @@ class DomainContentView(MethodView):
     def delete(self, domain_id):
         """Delete domain content."""
         domain = domain_manager.get(document_id=domain_id)
+        can_access = user_can_access_domain(domain)
+        if not can_access:
+            add_user_action(f"Delete Domain Content (ACCESS DENIED) - {domain['name']}")
+            return jsonify({"error": "User does not have permission to delete to this domain content"}), 400
+
 
         name = domain["name"]
         category = domain["category"]
@@ -201,6 +261,7 @@ class DomainContentView(MethodView):
         except requests.exceptions.HTTPError as e:
             return {"error": str(e)}, 400
 
+        add_user_action(f"Delete Domain Content - {domain['name']}")
         return jsonify(
             domain_manager.remove(
                 document_id=domain_id, data={"category": "", "s3_url": ""}
@@ -215,6 +276,11 @@ class DomainGenerateView(MethodView):
         """Create website."""
         category = request.args.get("category")
         domain = domain_manager.get(document_id=domain_id)
+        can_access = user_can_access_domain(domain)
+        if not can_access:
+            add_user_action(f"Generate domain content from template (ACCESS DENIED) - {domain['name']}")
+            return jsonify({"error": "User does not have permission to generate content for this domain"}), 400
+
 
         # Switch instance to unavailable to prevent user actions
         domain_manager.update(
@@ -252,6 +318,7 @@ class DomainGenerateView(MethodView):
                 },
             )
 
+            add_user_action(f"Generate domain content from template - Template: {category} - Website {domain['name']}")
             return jsonify(
                 {
                     "message": f"{domain_name} static site has been created from the {category} template."
@@ -273,10 +340,22 @@ class DomainRedirectView(MethodView):
 
     def get(self, domain_id):
         """Get all redirects for a domain."""
+        domain = domain_manager.get(document_id=domain_id)
+        can_access = user_can_access_domain(domain)
+        if not can_access:
+            add_user_action(f"Access redirects (ACCESS DENIED) - {domain['name']}")
+            return jsonify({"error": "User does not have permission to view redirects for this domain: {domain['name']}"}), 400
+
         return jsonify(domain_manager.get(document_id=domain_id, fields=["redirects"]))
 
     def post(self, domain_id):
         """Create a domain redirect."""
+        domain = domain_manager.get(document_id=domain_id)
+        can_access = user_can_access_domain(domain)
+        if not can_access:
+            add_user_action(f"Create redirect (ACCESS DENIED) - {domain['name']}")
+            return jsonify({"error": "User does not have permission to create redirects for this domain: {domain['name']}"}), 400
+
         data = {
             "subdomain": request.json["subdomain"],
             "redirect_url": request.json["redirect_url"],
@@ -296,6 +375,7 @@ class DomainRedirectView(MethodView):
             redirect_url=data["redirect_url"],
         )
 
+        add_user_action(f"Add redirect to domain - domain id: {domain_id} - redirect: data['redirect_url']")
         return jsonify(
             domain_manager.add_to_list(
                 document_id=domain_id, field="redirects", data=data
@@ -304,6 +384,12 @@ class DomainRedirectView(MethodView):
 
     def put(self, domain_id):
         """Update a subdomain redirect value."""
+        domain = domain_manager.get(document_id=domain_id)
+        can_access = user_can_access_domain(domain)
+        if not can_access:
+            add_user_action(f"Update redirects to domain (ACCESS DENIED) - {domain['name']}")
+            return jsonify({"error": "User does not have permission to update redirects for this domain: {domain['name']}"}), 400
+
         data = {
             "subdomain": request.json["subdomain"],
             "redirect_url": request.json["redirect_url"],
@@ -316,6 +402,8 @@ class DomainRedirectView(MethodView):
             subdomain=data["subdomain"],
             redirect_url=data["redirect_url"],
         )
+
+        add_user_action(f"Update redirect to domain - domain id: {domain_id} - redirect: data['redirect_url']")
         return jsonify(
             domain_manager.update_in_list(
                 document_id=domain_id,
@@ -327,10 +415,17 @@ class DomainRedirectView(MethodView):
 
     def delete(self, domain_id):
         """Delete a subdomain redirect."""
+        domain = domain_manager.get(document_id=domain_id)
+        can_access = user_can_access_domain(domain)
+        if not can_access:
+            add_user_action(f"Deelete redirects (ACCESS DENIED) - {domain['name']}")
+            return jsonify({"error": "User does not have permission to delete redirects for this domain: {domain['name']}"}), 400
+
         subdomain = request.args.get("subdomain")
         if not subdomain:
             return {"error": "must pass subdomain as a request arg to delete."}
         delete_redirect(domain_id=domain_id, subdomain=subdomain)
+        add_user_action(f"Delete redirect for domain - Domain id: {domain_id} - subdomain: {subdomain}")
         return jsonify(
             domain_manager.delete_from_list(
                 document_id=domain_id,
@@ -346,6 +441,11 @@ class DomainLaunchView(MethodView):
     def get(self, domain_id):
         """Launch a static site."""
         domain = domain_manager.get(document_id=domain_id)
+        can_access = user_can_access_domain(domain)
+        if not can_access:
+            add_user_action(f"Launch Domain (ACCESS DENIED) - {domain['name']}")
+            return jsonify({"error": "User does not have permission to launch this domain: {domain['name']}"}), 400
+
 
         # Switch instance to unavailable to prevent user actions
         domain_manager.update(
@@ -370,8 +470,10 @@ class DomainLaunchView(MethodView):
                 data=data,
             )
             name = domain["name"]
+            add_user_action(f"Launch Domain - {data['name']}")
             return jsonify({"success": f"{name} has been launched"})
         except Exception as e:
+            add_user_action(f"Launch Domain - FAILED {data['name']}")
             logger.exception(e)
             # Switch instance to unavailable to prevent user actions
             domain_manager.update(
@@ -385,6 +487,11 @@ class DomainLaunchView(MethodView):
     def delete(self, domain_id):
         """Stop a static site."""
         domain = domain_manager.get(document_id=domain_id)
+        can_access = user_can_access_domain(domain)
+        if not can_access:
+            add_user_action(f"Take down a Domain site (ACCESS DENIED) - {domain['name']}")
+            return jsonify({"error": "User does not have permission to take down this domain: {domain['name']}"}), 400
+
 
         # Switch instance to unavailable to prevent user actions
         domain_manager.update(
@@ -411,6 +518,7 @@ class DomainLaunchView(MethodView):
                 document_id=domain_id,
                 data={"acm": "", "cloudfront": ""},
             )
+            add_user_action(f"Take down a Domain site - {domain['name']}")
             return jsonify(resp)
         except Exception as e:
             logger.exception(e)
@@ -429,6 +537,12 @@ class DomainRecordView(MethodView):
 
     def get(self, domain_id):
         """Get the hosted zone records for a domain."""
+        domain = domain_manager.get(document_id=domain_id)
+        can_access = user_can_access_domain(domain)
+        if not can_access:
+            add_user_action(f"View domain hosted zone records (ACCESS DENIED) - {domain['name']}")
+            return jsonify({"error": "User does not have permission to view the hosted zone records: {domain['name']}"}), 400
+
         hosted_zone_id = domain_manager.get(document_id=domain_id, fields=["route53"])[
             "route53"
         ]["id"]
@@ -447,12 +561,18 @@ class DomainCategorizeView(MethodView):
         chrome_options.add_argument("--headless")
         domain = domain_manager.get(document_id=domain_id)
         domain_name = domain["name"]
-        if domain.get("is_categorized", None):
-            return {"error": f"{domain_name} has already been categorized."}
 
         category = category_manager.get(
             filter_data={"name": request.args.get("category", "").capitalize()}
         )
+
+        can_access = user_can_access_domain(domain)
+        if not can_access:
+            add_user_action(f"Categorize domain (ACCESS DENIED) - {domain['name']}")
+            return jsonify({"error": "User does not have permission to categorize the domain: {domain['name']}"}), 400
+
+        if domain.get("is_categorized", None):
+            return {"error": f"{domain_name} has already been categorized."}
 
         if not category:
             return {"error": "Category does not exist"}
@@ -508,6 +628,7 @@ class DomainCategorizeView(MethodView):
             document_id=domain_id,
             data={"is_category_submitted": is_category_submitted},
         )
+        add_user_action(f"Categorize domain - domain: {domain['name']} - category: {category}")
         return jsonify(
             {
                 "message": f"{domain_name} has been successfully submitted for categorization"
@@ -519,7 +640,7 @@ class DomainCheckView(MethodView):
     """DomainCategoryCheckView."""
 
     def update_submission(self, query, dicts):
-        """Search through existing submissions and check as categorized."""
+        """Search through existing submissions and check as categorized."""        
         next(
             item.update({"is_categorized": True})
             for item in dicts
@@ -531,6 +652,11 @@ class DomainCheckView(MethodView):
     def get(self, domain_id):
         """Check category for a domain."""
         domain = domain_manager.get(document_id=domain_id)
+        can_access = user_can_access_domain(domain)
+        if not can_access:
+            add_user_action(f"Check domain category (ACCESS DENIED) - {domain['name']}")
+            return jsonify({"error": "User does not have permission to check the category for the domain: {domain['name']}"}), 400
+
 
         if not domain.get("is_category_submitted", None):
             return jsonify(
